requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python>=4.8.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.4.0
pydantic-settings>=2.0.0
websockets>=11.0.0
python-dotenv>=1.0.0